+++ packages/fuselage-ui-kit/src/index.js
@@ -9,7 +9,6 @@ import {
   TextInput,
   InputBox,
   Box,
-  Text,
 } from '@rocket.chat/fuselage';
 import {
   uiKitMessage,
@@ -18,8 +17,11 @@ import {
   UiKitParserMessage,
   ELEMENT_TYPES,
   UiKitParserModal,
+  UiKitParserButtons,
+  uiKitText,
+  uiKitButtons,
+  UiKitParserText,
 } from '@rocket.chat/ui-kit';
-import { load } from 'svg2ttf/lib/svg';
 
 import { Section as SectionLayoutBlock } from './Section';
 import { Actions as ActionsLayoutBlock } from './Actions';
@@ -27,16 +29,18 @@ import { Input as InputLayoutBlock } from './Input';
 import { MessageImage, ModalImage } from './Image';
 import { StaticSelect, MultiStaticSelect } from './StaticSelect';
 import { Block } from './Block';
+import { Overflow } from './Overflow';
 
 export const defaultContext = {
-  action: (...args) => alert(JSON.stringify(args)),
+  action: (...args) => console.log(JSON.stringify(args)),
   state: console.log,
   appId: '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz',
 };
 
 export const kitContext = React.createContext(defaultContext);
 
-const useBlockContext = ({ blockId, actionId, appId }, context) => {
+const useBlockContext = ({ blockId, actionId, appId, initialValue }, context) => {
+  const [value, setValue] = useState(initialValue);
   const [loading, setLoading] = useState(false);
   const { action, appId: appIdFromContext, state } = useContext(kitContext);
   if ([BLOCK_CONTEXT.SECTION, BLOCK_CONTEXT.ACTION].includes(context)) {
@@ -47,7 +51,8 @@ const useBlockContext = ({ blockId, actionId, appId }, context) => {
     }];
   }
 
-  return [{ loading, setLoading }, async ({ target: { value } }) => {
+  return [{ loading, setLoading, value }, async ({ target: { value } }) => {
+    setValue(value);
     setLoading(true);
     await state({ blockId, appId, actionId, value });
     setLoading(false);
@@ -64,96 +69,153 @@ const getStyle = (style) => {
   }
 };
 
-class MessageParser extends UiKitParserMessage {
-  button(element, context) {
+
+function mrkdwn({ text/* , type = 'plain_text'*/ } = { text: '' }) {
+  return text;
+}
+
+function plainText({ text /* , type = 'plain_text'*/ } = { text: '' }) {
+  return text;
+}
+
+function text({ text /* , type = 'plain_text'*/ } = { text: '' }) {
+  return text;
+}
+
+class TextParser extends UiKitParserText {
+  mrkdwn(...args) { return mrkdwn(...args); }
+
+  plainText(...args) { return plainText(...args); }
+
+  text(...args) { return text(...args); }
+}
+
+class ButtonsParser extends UiKitParserButtons {
+  button(element, context, key) {
     const [{ loading }, action] = useBlockContext(element, context);
     return (
       <Button
+        key={key}
+        mod-mod-loading={loading}
+        {...getStyle(element.style)}
+        small={context === BLOCK_CONTEXT.SECTION}
+        data-group={element.groupId}
+        key={element.actionId}
+        children={this.plainText(element.text)}
+        onClick={action}
+        value={element.value}
+      />
+    );
+  }
+}
+
+
+class MessageParser extends UiKitParserMessage {
+  mrkdwn(...args) { return mrkdwn(...args); }
+
+  plainText(...args) { return plainText(...args); }
+
+  text(...args) { return text(...args); }
+
+  overflow(element, context) {
+    const [{ loading }, action] = useBlockContext(element, context);
+    return <Overflow loading={loading} {...element} onChange={action} parser={this}/>;
+  }
+
+  button(element, context, key) {
+    const [{ loading }, action] = useBlockContext(element, context);
+    return (
+      <Button
+        key={key}
         mod-mod-loading={loading}
         {...getStyle(element.style)}
         small
         data-group={element.groupId}
         key={element.actionId}
-        children={this.text(element.text)}
+        children={this.plainText(element.text)}
         onClick={action}
         value={element.value}
       />
     );
   }
 
-  divider() {
-    return <Divider />;
+  divider(_, __, key) {
+    return <Margins block={'x24'} key={key}> <Divider /> </Margins>;
   }
 
-  text({ text/* , type = 'plain_text'*/ } = { text: '' }) {
-    return text;
-  }
-
-  section(args) {
+  section(args, context, index) {
     return (
-      <SectionLayoutBlock {...args} parser={this} />
+      <SectionLayoutBlock key={index} {...args} parser={this} />
     );
   }
 
-  actions(args) {
+  actions(args, _, key) {
     return (
-      <ActionsLayoutBlock {...args} parser={this} />
+      <ActionsLayoutBlock {...args} key={key} parser={this} />
     );
   }
 
-  datePicker(element, context) {
-    const [{ loading }, action] = useBlockContext(element, context);
+  datePicker(element, context, key) {
+    const [{ loading, value }, action] = useBlockContext(element, context);
     const { actionId, placeholder } = element;
     return (
       <InputBox
+        key={key}
+        value={value}
         mod-mod-loading={loading}
         id={actionId}
         name={actionId}
         rows={6}
         onInput={action}
-        placeholder={this.text(placeholder)}
+        placeholder={this.plainText(placeholder)}
         type='date'
       />
     );
   }
 
-  image(element, context) {
-    return <MessageImage element={element} context={context}/>;
+  image(element, context, key) {
+    return <MessageImage key={key} element={element} context={context}/>;
   }
 
-  context({ elements }/* , context*/) {
+  context({ elements }, context, key) {
     return (
-      <Flex.Container alignItems='center'>
-        <Block>
-          <Box is='div'>
-            {elements.map((element) => (
-              <Margins all={4}>
-                <Flex.Item>
-                  {[
-                    ELEMENT_TYPES.PLAIN_TEXT_INPUT,
-                    ELEMENT_TYPES.MARKDOWN,
-                  ].includes(element.type) ? (
-                      <Box is='span' textStyle='caption'>
-                        {this.renderContext(element, BLOCK_CONTEXT.CONTEXT, this)}
-                      </Box>
-                    )
-                    : this.renderContext(element, BLOCK_CONTEXT.CONTEXT, this)
+      <Block key={key}>
+        <Box is='div'>
+          <Flex.Container alignItems='center'>
+            <Margins all='neg-x4'>
+              <Box is='div'>
+                {elements.map((element, i) => (
+                  <Margins all='x4' key={i}>
+                    <Flex.Item>
+                      {[
+                        ELEMENT_TYPES.PLAIN_TEXT,
+                        ELEMENT_TYPES.MARKDOWN,
+                      ].includes(element.type) ? (
+                          <Box is='span' textStyle='c1' textColor='info'>
+                            {this.renderContext(element, BLOCK_CONTEXT.CONTEXT, this)}
+                          </Box>
+                        )
+                        : this.renderContext(element, BLOCK_CONTEXT.CONTEXT, this)
                     || element.type
-                  }
-                </Flex.Item>
-              </Margins>
-            ))}
-          </Box>
-        </Block>
-      </Flex.Container>
+                      }
+                    </Flex.Item>
+                  </Margins>
+                ))}
+              </Box>
+            </Margins>
+          </Flex.Container>
+        </Box>
+      </Block>
     );
   }
 
-  multiStaticSelect(element, context) {
-    const [{ loading }, action] = useBlockContext(element, context);
+  multiStaticSelect(element, context, key) {
+    const [{ loading, value }, action] = useBlockContext(element, context);
     return (
       <MultiStaticSelect
         {...element}
+        key={key}
+        value={value}
         mod-loading={loading}
         onChange={action}
         parser={this}
@@ -161,15 +223,15 @@ class MessageParser extends UiKitParserMessage {
     );
   }
 
-  staticSelect(element, context) {
-    const [{ loading }, action] = useBlockContext(element, context);
-    return <StaticSelect {...element} mod-loading={loading} onChange={action} parser={this} />;
+  staticSelect(element, context, key) {
+    const [{ loading, value }, action] = useBlockContext(element, context);
+    return <StaticSelect value={value} key={key} {...element} mod-loading={loading} onChange={action} parser={this} />;
   }
 
-  selectInput(element, context) {
-    const [{ loading }, action] = useBlockContext(element, context);
+  selectInput(element, context, key) {
+    const [{ loading, value }, action] = useBlockContext(element, context);
     return (
-      <SelectInput onChange={action} mod-loading={loading} placeholder={element.type} disabled />
+      <SelectInput key={key} value={value} onChange={action} mod-loading={loading} placeholder={element.type} disabled />
     );
   }
 }
@@ -182,39 +244,47 @@ class ModalParser extends UiKitParserModal {
     });
   }
 
-  input({ element, label, blockId, appId }) {
+  input({ element, label, blockId, appId }, context, index) {
     return (
       <InputLayoutBlock
+        key={index}
+        index={index}
         parser={this}
         element={{ ...element, appId, blockId }}
-        label={this.text(label)}
+        label={this.plainText(label)}
       />
     );
   }
 
-  image(element, context) {
-    return <ModalImage element={element} context={context}/>;
+  image(element, context, index) {
+    return <ModalImage key={index} element={element} context={context}/>;
   }
 
-  plainInput(element, context) {
-    const [{ loading }, action] = useBlockContext(element, context);
+  plainInput(element, context, index) {
+    const [{ loading, value }, action] = useBlockContext(element, context);
     const { multiline, actionId, placeholder } = element;
     const Component = multiline ? TextAreaInput : TextInput;
     return (
       <Component
+        key={index}
         mod-loading={loading}
         id={actionId}
         name={actionId}
         rows={6}
+        value={value}
         onInput={action}
-        placeholder={this.text(placeholder)}
+        placeholder={this.plainText(placeholder)}
       />
     );
   }
 }
 
+export const textParser = new TextParser();
 export const messageParser = new MessageParser();
 export const modalParser = new ModalParser();
+export const buttonsParser = new ButtonsParser();
 
+export const UiKitButtons = uiKitButtons();
+export const UiKitText = uiKitText(textParser);
 export const UiKitMessage = uiKitMessage(messageParser);
 export const UiKitModal = uiKitModal(modalParser);